+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Throw an error when a document does not have a `loadQueryBuilderData` method while using the `FormApi#queryBuilderData` action
 * `QueryBuilderValue` instances will now always add a type to the `value_explicit` object
 * Make `al-button` elements submit forms automatically
+* Fix operator `al-select` element staying empty in a query builder entry
 
 ## 0.2.0 (2022-11-02)